+++ src/app/(user)/(service)/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/src/components/auth/login/login-form";
+import { LoginForm } from "@/src/features/auth/auth-login/components/login-form";
 
 export default function LoginPage() {
   return <LoginForm />;
