+++ src/app/(user)/(service)/(auth)/reset-password/verify/page.tsx
@@ -1,4 +1,4 @@
-import { ResetPasswordVerification } from "@/src/components/auth/reset-password/reset-password-verification";
+import { ResetPasswordVerification } from "@/src/features/auth/auth-reset-pw/components/reset-password-verification";
 
 export default function ResetPasswordVerifyPage() {
   return <ResetPasswordVerification />;
