+++ src/contexts/AuthProvider.tsx
@@ -5,6 +5,6 @@ type Props = {
   children: React.ReactNode;
 };
 
-export default function AuthContext({ children }: Props) {
+export default function AuthProvider({ children }: Props) {
   return <SessionProvider>{children}</SessionProvider>;
 }
