+++ src/features/auth/auth-common/components/auth-side-image.tsx
