+++ src/features/auth/auth-reset-pw/components/reset-password-verification.tsx
@@ -5,12 +5,12 @@ import Image from "next/image";
 import logo from "@/public/imgs/logo.png";
 import { useState } from "react";
 import { EmailForm } from "./email-form";
-import { OTPForm } from "./OTP-form";
 import { useRouter } from "next/navigation";
 import { sendAuthCode, verifyAuthCode } from "@/src/services/auth.service";
 import { toast } from "sonner";
 import axios from "axios";
 import dayjs from "dayjs";
+import { OTPForm } from "./otp-form";
 
 export function ResetPasswordVerification() {
   const router = useRouter();
